Restructure code base and ensure tests still pass
