+++ docs/user/reference/api.rst
@@ -27,7 +27,6 @@ This is the internal API reference for ophyd_async
     ophyd_async.core.backends
     ophyd_async.core.devices
     ophyd_async.core.epicsdemo
-    ophyd_async.core.signals
+    ophyd_async.core.signal
     ophyd_async.core.async_status
-    ophyd_async.core.device_collector
     ophyd_async.core.utils
