--- other_licenses/pyepics
@@ -1,84 +1,0 @@
-The epics python module was orignally written by
-
-   Matthew Newville <newville@cars.uchicago.edu>
-   CARS, University of Chicago
-
-There have been several contributions from many others, notably Angus
-Gratton <angus.gratton@anu.edu.au>.  See the Acknowledgements section of
-the documentation for a list of more contributors.
-
-Except where explicitly noted, all files in this distribution are licensed
-under the Epics Open License.:
-
-------------------------------------------------
-
-Copyright  2010  Matthew Newville, The University of Chicago.  All rights reserved.
-
-The epics python module is distributed subject to the following license conditions:
-SOFTWARE LICENSE AGREEMENT
-Software: epics python module
-
-   1.  The "Software", below, refers to the epics python module (in either
-   source code, or binary form and accompanying documentation). Each
-   licensee is addressed as "you" or "Licensee."
-
-   2.  The copyright holders shown above and their third-party licensors
-   hereby grant Licensee a royalty-free nonexclusive license, subject to
-   the limitations stated herein and U.S. Government license rights.
-
-   3.  You may modify and make a copy or copies of the Software for use
-   within your organization, if you meet the following conditions:
-
-       1. Copies in source code must include the copyright notice and  this
-       Software License Agreement.
-
-       2. Copies in binary form must include the copyright notice and  this
-       Software License Agreement in the documentation and/or other
-       materials provided with the copy.
-
-   4.  You may modify a copy or copies of the Software or any portion of
-   it, thus forming a work based on the Software, and distribute copies of
-   such work outside your organization, if you meet all of the following
-   conditions:
-
-       1. Copies in source code must include the copyright notice and this
-       Software License Agreement;
-
-       2. Copies in binary form must include the copyright notice and this
-       Software License Agreement in the documentation and/or other
-       materials provided with the copy;
-
-       3. Modified copies and works based on the Software must carry
-       prominent notices stating that you changed specified portions of
-       the Software.
-
-   5.  Portions of the Software resulted from work developed under a
-   U.S. Government contract and are subject to the following license: the
-   Government is granted for itself and others acting on its behalf a
-   paid-up, nonexclusive, irrevocable worldwide license in this computer
-   software to reproduce, prepare derivative works, and perform publicly
-   and display publicly.
-
-   6.  WARRANTY DISCLAIMER. THE SOFTWARE IS SUPPLIED "AS IS" WITHOUT
-   WARRANTY OF ANY KIND. THE COPYRIGHT HOLDERS, THEIR THIRD PARTY
-   LICENSORS, THE UNITED STATES, THE UNITED STATES DEPARTMENT OF ENERGY,
-   AND THEIR EMPLOYEES: (1) DISCLAIM ANY WARRANTIES, EXPRESS OR IMPLIED,
-   INCLUDING BUT NOT LIMITED TO ANY IMPLIED WARRANTIES OF MERCHANTABILITY,
-   FITNESS FOR A PARTICULAR PURPOSE, TITLE OR NON-INFRINGEMENT, (2) DO NOT
-   ASSUME ANY LEGAL LIABILITY OR RESPONSIBILITY FOR THE ACCURACY,
-   COMPLETENESS, OR USEFULNESS OF THE SOFTWARE, (3) DO NOT REPRESENT THAT
-   USE OF THE SOFTWARE WOULD NOT INFRINGE PRIVATELY OWNED RIGHTS, (4) DO
-   NOT WARRANT THAT THE SOFTWARE WILL FUNCTION UNINTERRUPTED, THAT IT IS
-   ERROR-FREE OR THAT ANY ERRORS WILL BE CORRECTED.
-
-   7.  LIMITATION OF LIABILITY. IN NO EVENT WILL THE COPYRIGHT HOLDERS,
-   THEIR THIRD PARTY LICENSORS, THE UNITED STATES, THE UNITED STATES
-   DEPARTMENT OF ENERGY, OR THEIR EMPLOYEES: BE LIABLE FOR ANY INDIRECT,
-   INCIDENTAL, CONSEQUENTIAL, SPECIAL OR PUNITIVE DAMAGES OF ANY KIND OR
-   NATURE, INCLUDING BUT NOT LIMITED TO LOSS OF PROFITS OR LOSS OF DATA,
-   FOR ANY REASON WHATSOEVER, WHETHER SUCH LIABILITY IS ASSERTED ON THE
-   BASIS OF CONTRACT, TORT (INCLUDING NEGLIGENCE OR STRICT LIABILITY), OR
-   OTHERWISE, EVEN IF ANY OF SAID PARTIES HAS BEEN WARNED OF THE
-   POSSIBILITY OF SUCH LOSS OR DAMAGES.
-
-------------------------------------------------
