+++ src/ophyd_async/core/devices/__init__.py
@@ -1,9 +1,11 @@
 from .device import Device, connect_children, get_device_children, name_children
+from .device_collector import DeviceCollector
 from .device_vector import DeviceVector
 from .standard_readable import StandardReadable
 
 __all__ = [
     "Device",
+    "DeviceCollector",
     "connect_children",
     "get_device_children",
     "name_children",
