+++ src/ophyd_async/core/epicsdemo/__init__.py
@@ -8,15 +8,8 @@ from typing import Callable, List, Optional
 import numpy as np
 from bluesky.protocols import Movable, Stoppable
 
-from ophyd_async.core import (
-    AsyncStatus,
-    Device,
-    StandardReadable,
-    epics_signal_r,
-    epics_signal_rw,
-    epics_signal_x,
-    observe_value,
-)
+from ophyd_async.core import AsyncStatus, Device, StandardReadable, observe_value
+from ophyd_async.epics.signal import epics_signal_r, epics_signal_rw, epics_signal_x
 
 
 class EnergyMode(Enum):
