+++ src/ophyd_async/core/signal.py
@@ -13,11 +13,11 @@ from bluesky.protocols import (
     Subscribable,
 )
 
-from ..async_status import AsyncStatus
-from ..backends.signal_backend import SignalBackend
-from ..backends.sim import SimSignalBackend
-from ..devices import Device
-from ..utils import DEFAULT_TIMEOUT, Callback, ReadingValueCallback, T
+from .async_status import AsyncStatus
+from .backends.signal_backend import SignalBackend
+from .backends.sim import SimSignalBackend
+from .devices import Device
+from .utils import DEFAULT_TIMEOUT, Callback, ReadingValueCallback, T
 
 _sim_backends: Dict[Signal, SimSignalBackend] = {}
