--- src/ophyd_async/core/signals/__init__.py
@@ -1,39 +1,0 @@
-from .epics import (
-    EpicsTransport,
-    epics_signal_r,
-    epics_signal_rw,
-    epics_signal_w,
-    epics_signal_x,
-)
-from .signal import (
-    Signal,
-    SignalR,
-    SignalRW,
-    SignalW,
-    SignalX,
-    observe_value,
-    set_and_wait_for_value,
-    set_sim_callback,
-    set_sim_put_proceeds,
-    set_sim_value,
-    wait_for_value,
-)
-
-__all__ = [
-    "EpicsTransport",
-    "epics_signal_r",
-    "epics_signal_rw",
-    "epics_signal_w",
-    "epics_signal_x",
-    "Signal",
-    "SignalR",
-    "SignalRW",
-    "SignalW",
-    "SignalX",
-    "observe_value",
-    "set_and_wait_for_value",
-    "set_sim_callback",
-    "set_sim_put_proceeds",
-    "set_sim_value",
-    "wait_for_value",
-]
