+++ src/ophyd_async/epics/areadetector/nd_plugin.py
@@ -1,0 +1,13 @@
+from ophyd_async.core.devices import Device
+
+from .utils import ad_r
+
+
+class NDPlugin(Device):
+    pass
+
+
+class NDPluginStats(NDPlugin):
+    def __init__(self, prefix: str) -> None:
+        # Define some signals
+        self.unique_id = ad_r(int, prefix + "UniqueId")
