+++ src/ophyd_async/epics/signal/__init__.py
@@ -1,0 +1,10 @@
+from .epics_transport import EpicsTransport
+from .signal import epics_signal_r, epics_signal_rw, epics_signal_w, epics_signal_x
+
+__all__ = [
+    "EpicsTransport",
+    "epics_signal_r",
+    "epics_signal_rw",
+    "epics_signal_w",
+    "epics_signal_x",
+]
