+++ tests/conftest.py
@@ -8,12 +8,14 @@ from typing import Any, Callable
 import pytest
 from bluesky.run_engine import RunEngine, TransitionError
 
-RECORD = str(Path(__file__).parent / "devices" / "db" / "panda.db")
+RECORD = str(Path(__file__).parent / "panda" / "db" / "panda.db")
 INCOMPLETE_BLOCK_RECORD = str(
-    Path(__file__).parent / "devices" / "db" / "incomplete_block_panda.db"
+    Path(__file__).parent / "panda" / "db" / "incomplete_block_panda.db"
 )
-INCOMPLETE_RECORD = str(Path(__file__).parent  / "devices" / "db" / "incomplete_panda.db")
-EXTRA_BLOCKS_RECORD = str(Path(__file__).parent / "devices" / "db" / "extra_blocks_panda.db")
+INCOMPLETE_RECORD = str(Path(__file__).parent / "panda" / "db" / "incomplete_panda.db")
+EXTRA_BLOCKS_RECORD = str(
+    Path(__file__).parent / "panda" / "db" / "extra_blocks_panda.db"
+)
 
 
 @pytest.fixture(scope="function")
@@ -84,3 +86,18 @@ async def failing_coroutine() -> Callable[[], Any]:
         raise ValueError()
 
     return inner_coroutine
+
+
+class DocHolder:
+    def __init__(self):
+        self.names = []
+        self.docs = []
+
+    def append(self, name, doc):
+        self.names.append(name)
+        self.docs.append(doc)
+
+
+@pytest.fixture
+def doc_holder() -> DocHolder:
+    return DocHolder()
