+++ tests/core/backends/test_sim.py
@@ -9,7 +9,7 @@ import pytest
 from bluesky.protocols import Reading
 
 from ophyd_async.core.backends import SignalBackend, SimSignalBackend
-from ophyd_async.core.signals import Signal
+from ophyd_async.core.signal import Signal
 from ophyd_async.core.utils import T
