+++ tests/epics/areadetector/test_hdf_streamer_det.py
@@ -6,69 +6,17 @@ import bluesky.plan_stubs as bps
 import bluesky.plans as bp
 import bluesky.preprocessors as bpp
 import pytest
-from ophyd_async.core.device_collector import DeviceCollector
-from ophyd_async.core.signals import set_sim_put_proceeds, set_sim_value
 
-from ophyd_async.devices.areadetector import (
+from ophyd_async.core.devices.device_collector import DeviceCollector
+from ophyd_async.core.signal import set_sim_put_proceeds, set_sim_value
+from ophyd_async.epics.areadetector import (
     ADDriver,
     FileWriteMode,
     HDFStreamerDet,
     ImageMode,
     NDFileHDF,
-    NDPluginStats,
-    SingleTriggerDet,
     TmpDirectoryProvider,
 )
-
-
-@pytest.fixture
-async def single_trigger_det():
-    async with DeviceCollector(sim=True):
-        stats = NDPluginStats("PREFIX:STATS")
-        det = SingleTriggerDet(
-            drv=ADDriver("PREFIX:DRV"), stats=stats, read_uncached=[stats.unique_id]
-        )
-
-    assert det.name == "det"
-    assert stats.name == "det-stats"
-    # Set non-default values to check they are set back
-    # These are using set_sim_value to simulate the backend IOC being setup
-    # in a particular way, rather than values being set by the Ophyd signals
-    set_sim_value(det.drv.acquire_time, 0.5)
-    set_sim_value(det.drv.array_counter, 1)
-    set_sim_value(det.drv.image_mode, ImageMode.continuous)
-    set_sim_value(stats.unique_id, 3)
-    yield det
-
-
-class DocHolder:
-    def __init__(self):
-        self.names = []
-        self.docs = []
-
-    def append(self, name, doc):
-        self.names.append(name)
-        self.docs.append(doc)
-
-
-async def test_single_trigger_det(single_trigger_det: SingleTriggerDet, RE):
-    d = DocHolder()
-    RE(bp.count([single_trigger_det]), d.append)
-
-    drv = single_trigger_det.drv
-    assert 1 == await drv.acquire.get_value()
-    assert ImageMode.single == await drv.image_mode.get_value()
-    assert True is await drv.wait_for_plugins.get_value()
-
-    assert d.names == ["start", "descriptor", "event", "stop"]
-    _, descriptor, event, _ = d.docs
-    assert descriptor["configuration"]["det"]["data"]["det-drv-acquire_time"] == 0.5
-    assert (
-        descriptor["data_keys"]["det-stats-unique_id"]["source"]
-        == "sim://PREFIX:STATSUniqueId_RBV"
-    )
-    assert event["data"]["det-drv-array_counter"] == 1
-    assert event["data"]["det-stats-unique_id"] == 3
 
 
 @pytest.fixture
@@ -103,9 +51,10 @@ async def hdf_streamer_dets():
     yield deta, detb
 
 
-async def test_hdf_streamer_dets_step(hdf_streamer_dets: List[HDFStreamerDet], RE):
-    d = DocHolder()
-    RE(bp.count(hdf_streamer_dets), d.append)
+async def test_hdf_streamer_dets_step(
+    hdf_streamer_dets: List[HDFStreamerDet], RE, doc_holder
+):
+    RE(bp.count(hdf_streamer_dets), doc_holder.append)  # type: ignore
 
     drv = hdf_streamer_dets[0].drv
     assert 1 == await drv.acquire.get_value()
@@ -118,7 +67,7 @@ async def test_hdf_streamer_dets_step(hdf_streamer_dets: List[HDFStreamerDet], R
     assert 0 == await hdf.num_capture.get_value()
     assert FileWriteMode.stream == await hdf.file_write_mode.get_value()
 
-    assert d.names == [
+    assert doc_holder.names == [  # type: ignore
         "start",
         "descriptor",
         "stream_resource",
@@ -128,7 +77,7 @@ async def test_hdf_streamer_dets_step(hdf_streamer_dets: List[HDFStreamerDet], R
         "event",
         "stop",
     ]
-    _, descriptor, sra, sda, srb, sdb, event, _ = d.docs
+    _, descriptor, sra, sda, srb, sdb, event, _ = doc_holder.docs  # type: ignore
     assert descriptor["configuration"]["deta"]["data"]["deta-drv-acquire_time"] == 0.8
     assert descriptor["configuration"]["detb"]["data"]["detb-drv-acquire_time"] == 1.8
     assert descriptor["data_keys"]["deta"]["shape"] == [768, 1024]
@@ -146,9 +95,8 @@ async def test_hdf_streamer_dets_step(hdf_streamer_dets: List[HDFStreamerDet], R
 # TODO: write test where they are in the same stream after
 #   https://github.com/bluesky/bluesky/issues/1558
 async def test_hdf_streamer_dets_fly_different_streams(
-    hdf_streamer_dets: List[HDFStreamerDet], RE
+    hdf_streamer_dets: List[HDFStreamerDet], RE, doc_holder
 ):
-    d = DocHolder()
     deta, detb = hdf_streamer_dets
 
     for det in hdf_streamer_dets:
@@ -174,10 +122,10 @@ async def test_hdf_streamer_dets_fly_different_streams(
                 yield from bps.collect(det, stream=True, return_payload=False)
         yield from bps.wait(group="complete")
 
-    RE(fly_det(5), d.append)
+    RE(fly_det(5), doc_holder.append)  # type: ignore
 
     # TODO: stream_* will come after descriptor soon
-    assert d.names == [
+    assert doc_holder.names == [  # type: ignore
         "start",
         "stream_resource",
         "stream_datum",
@@ -199,7 +147,7 @@ async def test_hdf_streamer_dets_fly_different_streams(
     assert 0 == await hdf.num_capture.get_value()
     assert FileWriteMode.stream == await hdf.file_write_mode.get_value()
 
-    _, sra, sda, descriptora, srb, sdb, descriptorb, _ = d.docs
+    _, sra, sda, descriptora, srb, sdb, descriptorb, _ = doc_holder.docs  # type: ignore
 
     assert descriptora["configuration"]["deta"]["data"]["deta-drv-acquire_time"] == 0.8
     assert descriptorb["configuration"]["detb"]["data"]["detb-drv-acquire_time"] == 1.8
@@ -220,7 +168,7 @@ async def test_hdf_streamer_dets_timeout(hdf_streamer_dets: List[HDFStreamerDet]
     set_sim_put_proceeds(det.drv.acquire, False)
     await det.kickoff()
     t = time.monotonic()
-    with patch("ophyd_async.devices.areadetector.FRAME_TIMEOUT", 0.1):
+    with patch("ophyd_async.epics.areadetector.hdf_streamer_det.FRAME_TIMEOUT", 0.1):
         with pytest.raises(TimeoutError, match="deta-hdf: writing stalled on frame 1"):
             await det.complete()
     assert 1.0 < time.monotonic() - t < 1.1
